+++ wnfs-wasm/tests/mock.ts
@@ -30,16 +30,20 @@ class MemoryBlockStore {
 
   /** Stores an array of bytes in the block store. */
   async getBlock(cid: Uint8Array): Promise<Uint8Array | undefined> {
-    const decoded_cid = CID.decode(cid);
-    return this.store.get(decoded_cid.toString());
+    const decodedCid = CID.decode(cid);
+    return this.store.get(decodedCid.toString());
   }
 
   /** Retrieves an array of bytes from the block store with given CID. */
-  async putBlock(bytes: Uint8Array, codec: number): Promise<Uint8Array> {
-    const hash = await sha256.digest(bytes);
-    const cid = CID.create(1, codec, hash);
-    this.store.set(cid.toString(), bytes);
-    return cid.bytes;
+  async putBlockKeyed(cid: Uint8Array, bytes: Uint8Array): Promise<void> {
+    const decodedCid = CID.decode(cid);
+    this.store.set(decodedCid.toString(), bytes);
+  }
+
+  /** Finds out whether a block is retrievable from this blockstore */
+  async hasBlock(cid: Uint8Array): Promise<boolean> {
+    const decodedCid = CID.decode(cid);
+    return this.store.has(decodedCid.toString());
   }
 }
 
@@ -182,10 +186,42 @@ const createRecipientExchangeRoot = async (
   return [key, rootDir];
 };
 
+class Sha256BlockStore {
+  private store: Map<string, Uint8Array>;
+
+  constructor() {
+    this.store = new Map();
+  }
+
+  async getBlock(cid: Uint8Array): Promise<Uint8Array | undefined> {
+    const decodedCid = CID.decode(cid);
+    return this.store.get(decodedCid.toString());
+  }
+
+  async putBlockKeyed(cid: Uint8Array, bytes: Uint8Array): Promise<void> {
+    const decodedCid = CID.decode(cid);
+    this.store.set(decodedCid.toString(), bytes);
+  }
+
+  async hasBlock(cid: Uint8Array): Promise<boolean> {
+    const decodedCid = CID.decode(cid);
+    return this.store.has(decodedCid.toString());
+  }
+
+  // We overwrite the putBlock method
+  async putBlock(bytes: Uint8Array, codec: number): Promise<Uint8Array> {
+    const hash = await sha256.digest(bytes);
+    const cid = CID.create(1, codec, hash);
+    this.store.set(cid.toString(), bytes);
+    return cid.bytes;
+  }
+}
+
 export {
   sampleCID,
   CID,
   MemoryBlockStore,
+  Sha256BlockStore,
   Rng,
   createSharerDir,
   createRecipientExchangeRoot,