+++ wnfs/src/public/link.rs
@@ -52,31 +52,25 @@ impl PublicLink {
 
     /// Gets the Cid stored in type. It attempts to get it from the store if it is not present in type.
     #[inline]
-    pub async fn resolve_cid(&self, store: &(impl BlockStore + ?Sized)) -> Result<Cid> {
+    pub async fn resolve_cid(&self, store: &impl BlockStore) -> Result<Cid> {
         self.0.resolve_cid(store).await
     }
 
     /// Gets the value stored in link. It attempts to get it from the store if it is not present in link.
     #[inline]
-    pub async fn resolve_value(&self, store: &(impl BlockStore + ?Sized)) -> Result<&PublicNode> {
+    pub async fn resolve_value(&self, store: &impl BlockStore) -> Result<&PublicNode> {
         self.0.resolve_value(store).await
     }
 
     /// Gets mut value stored in link. It attempts to get it from the store if it is not present in link.
     #[inline]
-    pub async fn resolve_value_mut(
-        &mut self,
-        store: &(impl BlockStore + ?Sized),
-    ) -> Result<&mut PublicNode> {
+    pub async fn resolve_value_mut(&mut self, store: &impl BlockStore) -> Result<&mut PublicNode> {
         self.0.resolve_value_mut(store).await
     }
 
     /// Gets an owned value from type. It attempts to it get from the store if it is not present in type.
     #[inline]
-    pub async fn resolve_owned_value(
-        self,
-        store: &(impl BlockStore + ?Sized),
-    ) -> Result<PublicNode> {
+    pub async fn resolve_owned_value(self, store: &impl BlockStore) -> Result<PublicNode> {
         self.0.resolve_owned_value(store).await
     }