+++ wnfs-hamt/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0 (2024-02-15)
+
+* Bumped minimal supported rust version to 1.75
+* Moved all traits to use the new ["return position impl trait in trait" feature](https://blog.rust-lang.org/2023/12/21/async-fn-rpit-in-traits.html)
+* Update to new `BlockStore` trait from wnfs-common
+* Evolve `BlockStore` trait ([#402](https://github.com/wnfs-wg/rs-wnfs/pull/402))
+
 ## 0.1.26 (2023-12-06)
 
 * Refactored the API to use a `Storable` trait instead of `AsyncSerialize` ([#378](https://github.com/wnfs-wg/rs-wnfs/pull/378))